Module parquet2::compression

Functionality to compress and decompress data according to the parquet specification

Structs

BrotliLevel

Represents a valid brotli compression level.

GzipLevel

Represents a valid gzip compression level.

ZstdLevel

Represents a valid zstd compression level.

Enums

Compression

CompressionOptions

Defines the compression settings for writing a parquet file.

Functions

compress

Compresses data stored in slice input_buf and writes the compressed result to output_buf. Note that you’ll need to call clear() before reusing the same output_buf across different compress calls.

decompress

Decompresses data stored in slice input_buf and writes output to output_buf. Returns the total number of bytes written.