Trait parquet2::deserialize::HybridRleRunsIterator

pub trait HybridRleRunsIterator<'a>: Iterator<Item = Result<HybridEncoded<'a>, Error>> {
    // Required method
    fn number_of_elements(&self) -> usize;
}

Required Methods

fn number_of_elements(&self) -> usize

Number of elements remaining. This may not be the items of the iterator - an item of the iterator may contain more than one element.