Function parquet2::encoding::bitpacked::encode_pack

pub fn encode_pack<T: Unpackable>(
    unpacked: &[T],
    num_bits: usize,
    packed: &mut [u8]
)

Encodes (packs) a potentially incomplete pack of Unpackable into bitpacked bytes packed, using num_bits per value.

This function assumes that the maximum value in unpacked fits in num_bits bits and saturates higher values.

Only the first ceil8(unpacked.len() * num_bits) of packed are populated.