Function parquet2::encoding::delta_length_byte_array::encode

pub fn encode<A: AsRef<[u8]>, I: Iterator<Item = A> + Clone>(
    iterator: I, 
    buffer: &mut Vec<u8>
)

Encodes a clonable iterator of &[u8] into buffer. This does not allocated on the heap.

Implementation

This encoding is equivalent to call delta_bitpacked::encode on the lengths of the items of the iterator followed by extending the buffer from each item of the iterator.