Module parquet2::encoding::hybrid_rle

Structs

BitmapIter

An iterator of bits according to the LSB format

Decoder

An iterator that, given a slice of bytes, returns HybridEncoded

HybridRleDecoder

Iterator of u32 from a byte slice of Hybrid-RLE encoded values

Enums

HybridEncoded

The two possible states of an RLE-encoded run.

Functions

bitpacked_encode

Writes an iterator of bools into writer, with LSB first.

encode_bool

the bitpacked part of the encoder.

encode_u32

RLE-hybrid encoding of u32. This currently only yields bitpacked values.