Struct parquet2::page::DataPageHeaderV2

pub struct DataPageHeaderV2 {
    pub num_values: i32,
    pub num_nulls: i32,
    pub num_rows: i32,
    pub encoding: Encoding,
    pub definition_levels_byte_length: i32,
    pub repetition_levels_byte_length: i32,
    pub is_compressed: Option<bool>,
    pub statistics: Option<Statistics>,
}

New page format allowing reading levels without decompressing the data Repetition and definition levels are uncompressed The remaining section containing the data is compressed if is_compressed is true

Fields
num_values: i32

Number of values, including NULLs, in this data page. *

num_nulls: i32

Number of NULL values, in this data page. Number of non-null = num_values - num_nulls which is also the number of values in the data section *

num_rows: i32

Number of rows in this data page. which means pages change on record boundaries (r = 0) *

encoding: Encoding

Encoding used for data in this page *

definition_levels_byte_length: i32

length of the definition levels

repetition_levels_byte_length: i32

length of the repetition levels

is_compressed: Option<bool>

whether the values are compressed. Which means the section of the page between definition_levels_byte_length + repetition_levels_byte_length + 1 and compressed_page_size (included) is compressed with the compression_codec. If missing it is considered compressed

statistics: Option<Statistics>

optional statistics for the data in this page *

Implementations

impl DataPageHeaderV2

pub fn new<F7, F8>(
num_values: i32,
num_nulls: i32,
num_rows: i32,
encoding: Encoding,
definition_levels_byte_length: i32,
repetition_levels_byte_length: i32,
is_compressed: F7,
statistics: F8
) -> DataPageHeaderV2where
F7: Into<Option<bool>>,
F8: Into<Option<Statistics>>,

pub fn read_from_in_protocol<T>(
i_prot: &mut T
) -> Result<DataPageHeaderV2, Error>where
T: TInputProtocol,

pub async fn stream_from_in_protocol<T>(
i_prot: &mut T
) -> impl Future<Output = Result<DataPageHeaderV2, Error>>where
T: TInputStreamProtocol,

pub fn write_to_out_protocol<T>(&self, o_prot: &mut T) -> Result<usize, Error>where
T: TOutputProtocol,

pub async fn write_to_out_stream_protocol<T>(
&self,
o_prot: &mut T
) -> impl Future<Output = Result<usize, Error>>where
T: TOutputStreamProtocol,