Module parquet2::compression

Structs

BrotliCodec
Codec for Brotli compression algorithm.

GZipCodec
Codec for GZIP compression algorithm.

Lz4Codec
Codec for LZ4 compression algorithm.

SnappyCodec
Codec for Snappy compression format.

ZstdCodec
Codec for Zstandard compression algorithm.

Enums

Compression

Traits

Codec
Parquet compression codec interface.

Functions

create_codec
Given the compression type codec, returns a codec used to compress and decompress bytes for the compression type. This returns None if the codec type is UNCOMPRESSED.