Module parquet2::metadata

Structs

ColumnChunkMetaData

Metadata for a column chunk.

ColumnDescriptor

A descriptor for leaf-level primitive columns. This encapsulates information such as definition and repetition levels and is used to re-assemble nested data.

FileMetaData

Metadata for a Parquet file.

RowGroupMetaData

Metadata for a row group.

SchemaDescriptor

A schema descriptor. This encapsulates the top-level schemas for all the columns, as well as all descriptors for all the primitive columns.

Enums

ColumnOrder

Column order that specifies what method was used to aggregate min/max values for statistics.

SortOrder

Sort order for page and column statistics.

Functions

get_sort_order

Returns sort order for a physical/logical type.

Type Definitions

KeyValue