Module parquet2::schema::types

Structs

BasicTypeInfo

Common type information.

DecimalType

Decimal logical type annotation

IntType

Integer logical type annotation

TimeType

Time logical type annotation

TimestampType

Timestamp logical type annotation

Type

Types supported by Parquet. These types are intended to be used in combination with the encodings to control the on disk storage format. For example INT16 is not included as a type since a good encoding of INT32 would handle this.

Enums

GroupConvertedType

LogicalType

ParquetType

Representation of a Parquet type. Used to describe primitive leaf fields and structs, including top-level schema. Note that the top-level schema type is represented using GroupType whose repetition is None.

PhysicalType

PrimitiveConvertedType

TimeUnit

Functions

converted_to_group_converted

converted_to_primitive_converted

group_converted_converted_to

physical_type_to_type

primitive_converted_to_converted

type_to_physical_type